requirements.txt
streamlit
pillow
numpy
requests
opencv-python-headless
